tools/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile tools/requirements.in
#
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
billiard==4.2.1
    # via celery
celery[redis]==5.5.3
    # via -r tools/requirements.in
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-default-group
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   shot-scraper
    #   typer
click-default-group==1.2.4
    # via shot-scraper
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
dateutils==0.6.12
    # via -r tools/requirements.in
greenlet==3.2.3
    # via playwright
idna==3.10
    # via requests
inflection==0.5.1
    # via -r tools/requirements.in
kombu[redis]==5.5.4
    # via celery
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
packaging==25.0
    # via kombu
playwright==1.54.0
    # via
    #   -r tools/requirements.in
    #   shot-scraper
prompt-toolkit==3.0.51
    # via click-repl
pydantic==2.11.7
    # via -r tools/requirements.in
pydantic-core==2.33.2
    # via pydantic
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   dateutils
python-frontmatter==1.1.0
    # via -r tools/requirements.in
python-slugify==8.0.4
    # via -r tools/requirements.in
pytz==2025.2
    # via dateutils
pyyaml==6.0.2
    # via
    #   python-frontmatter
    #   shot-scraper
redis==5.2.1
    # via kombu
requests==2.32.4
    # via -r tools/requirements.in
rich==14.0.0
    # via
    #   -r tools/requirements.in
    #   typer
shellingham==1.5.4
    # via typer
shot-scraper==1.8
    # via -r tools/requirements.in
six==1.17.0
    # via python-dateutil
text-unidecode==1.3
    # via python-slugify
typer==0.16.0
    # via -r tools/requirements.in
typing-extensions==4.14.1
    # via
    #   pydantic
    #   pydantic-core
    #   pyee
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via kombu
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
